pyproject.toml
[tool.poetry]
name = "agentops-assistant"
version = "0.1.0"
description = ""
authors = ["Gopi Krishna <gopigof@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.12,<3.13"


[tool.poetry.group.pipeline.dependencies]
jupyter = "^1.1.1"
langchain-core = "^0.3.21"
pytesseract = "^0.3.13"
atlassian-python-api = "^3.41.16"
langchain-community = "^0.3.8"
lxml = "^5.3.0"
pdf2image = "^1.17.0"
langchain-openai = "^0.2.10"
langchain-pinecone = "^0.2.0"
markdownify = "^0.14.1"


[tool.poetry.group.backend.dependencies]
fastapi = "^0.115.6"
langchain-openai = "^0.2.11"
langchain-community = "^0.3.9"
langchain-pinecone = "^0.2.0"
pydantic = {extras = ["email"], version = "^2.10.3"}
bcrypt = "^4.2.1"
langgraph = "^0.2.56"
tavily-python = "^0.5.0"
python-jose = "^3.3.0"
boto3 = "^1.35.76"
passlib = "^1.7.4"
uvicorn = "^0.32.1"
snowflake-connector-python = "^3.12.4"
snowflake-sqlalchemy = "^1.7.1"


[tool.poetry.group.frontend.dependencies]
streamlit = "^1.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

